refactor(agent/agentmcp): extract coder-related tools to mcptools package
